requirements.txt
absl-py==1.0.0
actionlib==1.13.2
angles==1.9.13
appdirs==1.4.3
apturl==0.5.2
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
autobahn==17.10.1
Automat==0.8.0
autopy==4.0.0
AwesomeTkinter==2021.11.8
bcrypt==3.1.7
beautifulsoup4==4.11.1
bidict==0.22.0
blinker==1.4
bondpy==1.8.6
breezy==3.0.2
Brlapi==0.7.0
Brotli==1.0.9
bs4==0.0.1
cachetools==5.2.0
camera-calibration==1.16.0
camera-calibration-parsers==1.12.0
cassandra-pylib==0.0.0
catkin==0.8.10
catkin-pkg==0.5.2
catkin-pkg-modules==0.5.2
cbor==1.0.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
chrome-gnome-shell==0.0.0
chromedriver-autoinstaller==0.3.1
click==8.1.3
colorama==0.4.1
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
controller-manager==0.19.5
controller-manager-msgs==0.19.5
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.16.0
cycler==0.11.0
Cython==0.29.14
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.1
docutils==0.16
dulwich==0.19.15
duplicity==0.8.12.0
dynamic-reconfigure==1.7.2
empy==3.3.2
entrypoints==0.3
et-xmlfile==1.1.0
eventlet==0.33.1
fake-useragent==0.1.11
Faker==13.3.2
fasteners==0.14.1
fastimport==0.9.8
filelock==3.0.12
FireDM==2022.4.14
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SocketIO==5.2.0
flatbuffers==1.12
fonttools==4.30.0
future==0.18.2
gast==0.4.0
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.15
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpg===1.13.1-unknown
greenlet==1.1.2
grpcio==1.47.0
gunicorn==20.1.0
h11==0.13.0
h5py==3.7.0
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
image-geometry==1.16.0
importlib-metadata==4.12.0
imutils==0.5.4
incremental==16.10.1
interactive-markers==1.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
keras==2.9.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.0
language-selector==0.1
laser-geometry==1.6.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==14.0.6
lockfile==0.12.2
louis==3.12.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.1
mediapipe==0.8.9.1
message-filters==1.15.14
monotonic==1.5
more-itertools==4.2.0
mpi4py==3.0.3
mutagen==1.45.1
netifaces==0.10.4
nose==1.3.7
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
openpyxl==3.0.9
opt-einsum==3.3.0
outcome==1.1.0
ovs==2.13.5
packaging==20.9
pandas==1.4.1
paramiko==2.6.0
persepolis==3.2.0
pexpect==4.6.0
phonenumbers==8.10.2
Pillow==7.0.0
plyer==2.0.0
protobuf==3.20.0
psutil==5.5.1
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.21
pycryptodomex==3.6.1
pycups==1.9.73
pycurl==7.45.1
pydot==1.4.1
PyGithub==1.43.7
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyngrok==5.1.0
PyOpenGL==3.1.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
pyshine==0.0.9
PySocks==1.7.1
pystray==0.19.3
python-apt==2.0.0+ubuntu0.20.4.7
python-bidi==0.4.2
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-engineio==4.3.4
python-gitlab==2.0.1
python-gnupg==0.4.5
python-nmap==0.7.1
python-qt-binding==0.4.4
python-snappy==0.5.3
python-socketio==5.7.1
python-xlib==0.31
PyTrie==0.2
pytz==2021.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
reportlab==3.5.34
requests==2.21.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
resource-retriever==1.12.7
roman==2.0.0
rosbag==1.15.14
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosdep==0.21.0
rosdep-modules==0.21.0
rosdistro==0.9.0
rosdistro-modules==0.9.0
rosgraph==1.15.14
rosinstall==0.7.8
rosinstall-generator==0.1.22
roslaunch==1.15.14
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.14
rosmake==1.15.8
rosmaster==1.15.14
rosmsg==1.15.14
rosnode==1.15.14
rosparam==1.15.14
rospkg==1.4.0
rospkg-modules==1.4.0
rospy==1.15.14
rosservice==1.15.14
rostest==1.15.14
rostopic==1.15.14
rosunit==1.15.8
roswtf==1.15.14
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.11
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.16
rqt-launch==0.4.9
rqt-logger-level==0.4.11
rqt-moveit==0.5.10
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.9
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.3
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rsa==4.9
rviz==1.14.14
SecretStorage==2.3.1
selenium==4.1.3
sensor-msgs==1.13.1
service-identity==18.1.0
setproctitle==1.1.10
simple-websocket==0.8.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
sniffio==1.2.0
sortedcontainers==2.4.0
sounddevice==0.4.4
soupsieve==2.3.2.post1
systemd-python==234
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflowjs==3.19.0
termcolor==1.1.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
tk==0.1.0
topic-tools==1.15.14
trio==0.20.0
trio-websocket==0.9.2
Twisted==18.9.0
txaio==2.10.0
typing-extensions==4.3.0
u-msgpack-python==2.1
ua-parser==0.10.0
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.2
usb-creator==0.3.7
user-agents==2.2.0
vcstools==0.1.42
virtualenv==20.0.17
wadllib==1.3.3
webdriver-manager==3.5.4
webencodings==0.5.1
websockets==10.3
Werkzeug==2.2.2
wrapt==1.11.2
wsaccel==0.6.2
wsproto==1.1.0
wstool==0.1.18
xacro==1.14.13
xkit==0.0.0
youtube-dl==2020.3.24
yt-dlp==2022.4.8
zipp==3.8.1
zope.interface==4.7.1
